feat: use view model for home view
