+++ pubspec.yaml
@@ -54,6 +54,7 @@ dependencies:
   async: ^2.11.0
   path: ^1.8.3
   mqtt_client: ^10.0.0
+  provider: ^6.0.5
 
 dev_dependencies:
   flutter_test:
